requirements.txt
# ---------- Core ----------
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.10,<3.0
websockets>=14.0,<15.0

# ---------- Database ----------
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
redis>=5.2,<6.0

# ---------- Data ----------
pandas>=2.2,<3.0
numpy>=2.2,<3.0
yfinance>=0.2,<1.0
scipy>=1.15,<2.0

# ---------- ML ----------
lightgbm>=4.5,<5.0
xgboost>=2.1,<3.0
scikit-learn>=1.6,<2.0
optuna>=4.2,<5.0
shap>=0.46,<1.0
torch>=2.6,<3.0

# ---------- Monitoring ----------
prometheus-client>=0.21,<1.0
mlflow>=2.20,<3.0
sentry-sdk[fastapi]>=2.19,<3.0

# ---------- Task Queue ----------
celery[redis]>=5.4,<6.0

# ---------- HTTP ----------
requests>=2.32,<3.0

# ---------- Broker ----------
smartapi-python>=1.4,<2.0
pyotp>=2.9,<3.0
python-dotenv>=1.0,<2.0

# ---------- Testing ----------
pytest>=8.3,<9.0
httpx>=0.28,<1.0
pytest-asyncio>=0.25,<1.0

# ---------- Linting ----------
flake8>=7.1,<8.0
ruff>=0.9,<1.0
